Struct travis_after_all::Build

pub struct Build {
    // some fields omitted
}

The information of a full build

Methods

impl Build

fn from_env() -> Result<Build, Error>

Fetch the relevant information from the environment

This reads the environment variables TRAVIS_BUILD_ID and TRAVIS_JOB_NUMBER, which are automatically set by Travis.

It also reads the variable POLLING_INTERVAL with a default of 5. Set it to a higher value for a longer timeout.

fn is_leader(&self) -> bool

Whether or not the current environment is the build leader

fn build_matrix(&self) -> Result<Matrix, Error>

Fetch the build matrix for the current build

fn wait_for_others(&self) -> Result<(), Error>

Wait for all non-leader jobs to finish

Returns an Error::BuildNotFound if this build is not known to Travis. Returns an Error::FailedBuilds if at least one non-leader build failed.

This loops until it fails or succeeds, there is no way to exit the loop.